chore: screen logic to helper class, use comms
